organize better
